DRF/djrest/requirements.txt
ase==3.24.0
asgiref==3.8.1
certifi==2024.12.14
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
Django==5.1.3
django-crontab==0.7.1
django-elasticsearch-dsl==8.0
django-elasticsearch-dsl-drf==0.22.5
django-nine==0.2.7
django-prometheus==2.3.1
djangorestframework==3.15.2
elastic==5.2.5.3
elastic-transport==8.17.0
elasticsearch==8.17.0
elasticsearch-dsl==8.17.1
Faker==33.3.1
fonttools==4.55.3
kiwisolver==1.4.8
matplotlib==3.10.0
numpy==2.2.1
packaging==24.2
pillow==11.1.0
prometheus_client==0.21.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
scipy==1.15.1
six==1.17.0
spglib==2.5.0
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
